User delete

+++ app/Http/Controllers/API/PrimaryUserController.php
@@ -214,29 +214,45 @@ class PrimaryUserController extends BaseController
 
 
     public function delete_primary_user($user_id){
-      if(!empty($user_id))  {
-          $user = User::where('id','=',$user_id)->first();
-          if(!empty($user)){
-              if($user->delete()){
-                  redirect('/primary_user')->with(['deleteMsg'=>'User has been deleted successfully!']);
-              }
-          }
-      }
+        try{
+            if(!empty($user_id))  {
+                $user = User::where('id','=',$user_id)->first();
+                if(!empty($user)){
+                    if($user->delete()){
+                        return redirect('/primary_user')->with(['deleteMsg'=>'User has been deleted successfully!']);
+                    }else{
+                        echo "User not deleted";
+                    }
+                }
+            }
+        }catch(Exception $ex){
+            echo "User not deleted" + $ex.getMessage();
+        }
+     
     }
 
     public function delete_beneficiary_user($user_id){
-      if(!empty($user_id))  {
-          $user = User::where('id','=',$user_id)->first();
-          if(!empty($user)){
-              $user_email = $user->email;
-              if($user->delete()){
-                  $beneficiary_user = BeneficiaryUser::where('email','=',$user_email
-                  )->first();
-                  $beneficiary_user->delete();
-                  redirect('/primary_user')->with(['deleteMsg'=>'User has been deleted successfully!']);
-              }
-          }
-      }
+        try{
+            if(!empty($user_id))  {
+                $user = User::where('id','=',$user_id)->first();
+                if(!empty($user)){
+                    $user_email = $user->email;
+                    if($user->delete()){
+                        $beneficiary_user = BeneficiaryUser::where('email','=',$user_email
+                        )->first();
+                        if(!empty($beneficiary_user)){
+                            $beneficiary_user->delete();
+                        }else{
+                            echo "User not deleted!";
+                        }
+                        
+                        return redirect('/beneficiary_user')->with(['deleteMsg'=>'User has been deleted successfully!']);
+                    }
+                }
+            }
+        }catch(Exception $ex){
+            echo "User not deleted" + $ex.getMessage();
+        }
     }
 
 }
